avoid removing `use` statements in one edge case.
The edge case is when the consumption of the word does not directly
match the module name, nor the imported word.

This is the case:

    use Foo qw(baz);
    print Foo::baz(42);

The imported "baz" is not directly used, so that one can be removed.
But the "Foo" module cannot be removed.

Such code can be transformed into:

    use Foo qw();
    print Foo::baz(42);